client-projects/garage-organizer/wave3_requirements.txt
# Wave 3 Data Collection Requirements

# Reddit
praw==7.7.1

# YouTube
google-api-python-client==2.108.0
yt-dlp==2023.12.30

# Video transcription
openai-whisper==20231117
ffmpeg-python==0.2.0

# Video analysis
opencv-python==4.8.1.78
openai==1.6.1

# General
requests==2.31.0
python-dotenv==1.0.0
